✨ Refactor RPC API: move lifespan management to a dedicated events module and add unit tests for access rights
