+++ services/catalog/src/simcore_service_catalog/api/rpc/events.py
