+++ services/catalog/src/simcore_service_catalog/core/events.py
@@ -12,7 +12,7 @@ from servicelib.fastapi.prometheus_instrumentation import (
 )
 
 from .._meta import APP_FINISHED_BANNER_MSG, APP_STARTED_BANNER_MSG
-from ..api.rpc.routes import rpc_api_lifespan
+from ..api.rpc.events import rpc_api_lifespan
 from ..repository.events import database_lifespan
 from ..service.director import director_lifespan
 from ..service.function_services import function_services_lifespan
